Gemfileの編集

+++ db/migrate/20170804062411_add_lat_to_advices.rb
@@ -1,5 +1,5 @@
 class AddLatToAdvices < ActiveRecord::Migration[5.1]
   def change
-    add_column :advices, :lat, :double
+    add_column :advices, :lat, :float
   end
 end
